Break tests for handler into their own file
This just helps reduce the size of files and simplify test organisation.
